generator/requirements.txt
# generator/requirements.txt
kafka-python==2.0.2
